requirements.txt
fastapi
uvicorn
sentence-transformers
requests
python-dotenv
